+++ src/eventing/exceptions.py
@@ -1,2 +1,10 @@
 class NoEventLoopError(Exception):
     pass
+
+
+class NotAClassError(Exception):
+    pass
+
+
+class MissingClassMethodHandler(Exception):
+    pass