+++ tests/test_on.py
@@ -16,13 +16,10 @@ def test_listener_with_on_decorator_is_added_to_listeners(ee):
     assert ee.listeners("foo") == [listener]
 
 
-METHODS_ON_CLASS_SKIP_REASON = "Wishful functionality; might not be a good idea"
-
-
-@pytest.mark.skip(METHODS_ON_CLASS_SKIP_REASON)
 def test_instancemethod_decorated_with_on_isnt_added_until_instantiated(ee):
+    @ee.handle_methods
     class Foo:
-        @ee.on("foo")
+        @ee.on("foo", method=True)
         def listener(self):
             pass
 
@@ -31,12 +28,12 @@ def test_instancemethod_decorated_with_on_isnt_added_until_instantiated(ee):
     assert ee.listeners("foo") == [foo_instance.listener]
 
 
-@pytest.mark.skip(METHODS_ON_CLASS_SKIP_REASON)
 def test_instancemethod_decorated_with_on_is_correctly_bound_when_called(ee):
     called_arg = None
 
+    @ee.handle_methods
     class Foo:
-        @ee.on("foo")
+        @ee.on("foo", method=True)
         def listener(self):
             nonlocal called_arg
             called_arg = self
@@ -46,13 +43,13 @@ def test_instancemethod_decorated_with_on_is_correctly_bound_when_called(ee):
     assert called_arg == foo_instance
 
 
-@pytest.mark.skip(METHODS_ON_CLASS_SKIP_REASON)
 def test_classmethod_decorated_with_on_is_correctly_bound_when_called(ee):
     called_arg = None
 
+    @ee.handle_methods
     class Foo:
         @classmethod
-        @ee.on("foo")
+        @ee.on("foo", method=True)
         def listener(cls):
             nonlocal called_arg
             called_arg = cls
@@ -62,13 +59,13 @@ def test_classmethod_decorated_with_on_is_correctly_bound_when_called(ee):
     assert called_arg == Foo
 
 
-@pytest.mark.skip(METHODS_ON_CLASS_SKIP_REASON)
 def test_staticmethod_decorated_with_on_is_correctly_called(ee):
     called = False
 
+    @ee.handle_methods
     class Foo:
         @staticmethod
-        @ee.on("foo")
+        @ee.on("foo", method=True)
         def listener():
             nonlocal called
             called = True
@@ -76,3 +73,28 @@ def test_staticmethod_decorated_with_on_is_correctly_called(ee):
     assert ee.listeners("foo") == [Foo.listener]
     ee.emit("foo")
     assert called
+
+
+def test_should_warn_if_class_is_wrapped_without_method_listeners(ee):
+    with pytest.warns(UserWarning):
+
+        @ee.handle_methods
+        class Foo:
+            pass
+
+
+def test_should_error_if_method_listener_by_no_class_deco(ee):
+    with pytest.raises(eventing.MissingClassMethodHandler):
+
+        class Foo:
+            @ee.on("foo", method=True)
+            def listener(self):
+                pass
+
+
+def test_should_error_if_method_listener_without_class(ee):
+    with pytest.raises(ValueError):
+
+        @ee.on("foo", method=True)
+        def listener():
+            pass